🐛 Fix: Comments
comments 수정합니다.

by @ehfdl
